Simple js type to string wip (#1)
* Very cheesy simple JS mode for typeToString

The method of passing flags is brutish and clunky, but not actually all
that hacky.

Works in Javascript files for (1) .types test output (2) language service.

* Short test of Javascript-simplified typeToString.

Probably not needed given the other coverage we have. But it's nice to
have a scratchpad for now.

* More typeToString JS simplifications

1. Simplify A & B -> object
2. Simplify keyof K -> string
3. Simplify T[K] -> object
4. Simplify T -> constraint of T, or object if no constraint

(4) doesn't work properly yet. I don't know if it shows the constraint,
and it definitely doesn't work for types where the parameter is nested,
like T[] or [T, U].

* Update new test's baselines

* Add more test cases for simplified typeToString

* Pass a `simple` flag into from server, return simple+complex

Not tested yet, but the basic design is for the server to return
an extra `simpleDisplayParts` property when appropriate so that the
consumer of the API can display that if it wants to.

* Only simplify when asked by the language service

Also scribbling on the language service, trying to temporarily default
javascript to simplified version.

* Server: boolean parameter;services:bool->flag;checker:flag

Previously it was a bit of a mess, but now the server passes a boolean
to services. False means 'provide just displayString', true
means 'provide displayString and simpleDisplayString'.

The services layer passes the boolean around, plus uses it to decide
whether to call the compiler in simple mode.

The compiler uses TypeFormatFlags.SuperSimple exclusively to decide whether to
simplify. But it never introduces SuperSimple; the services needs to
provide every time it calls into the compiler.

Note: It's now apparent that the compiler is a bit wrong. For example,
it converts almost every *symbol* to object, which is not correct.

* Make test harness simplify types in .js files

This may just be a temporary change

* Fix simplification of type parameters

And also returning simpleDisplayParts

* Make simpleJsTypes test even more obvious

* Update baselines

* Cleanup to make diff smaller

* Update more baselines

Not sure how these got missed earlier.

* Even more cleanup